Extract fullscreen logic to dedicated store
- Move fullscreen state and toggle function to `useFullscreenStore`
- Centralize fullscreen management and event listeners for better reusability
- Add a fullscreen logo and exit button within the background component
