+++ src/app.jsx
@@ -1,9 +1,10 @@
 import { Router, A } from "@solidjs/router";
 import { FileRoutes } from "@solidjs/start/router";
-import { Suspense, createSignal, onMount, onCleanup } from "solid-js";
+import { Suspense } from "solid-js";
 import { clientOnly } from "@solidjs/start";
 import { useThemeStore } from "~/stores/theme";
 import { useAnimationStore } from "~/stores/animation";
+import { useFullscreenStore } from "~/stores/fullscreen";
 import { BsFullscreen } from '@aminya/solid-icons/bs/BsFullscreen';
 import { BsFullscreenExit } from '@aminya/solid-icons/bs/BsFullscreenExit';
 import TopMenu from "~/components/TopMenu";
@@ -22,44 +23,7 @@ const ClientAnimationModeToggle = clientOnly(() =>
 );
 
 export default function App() {
-  const [isFullscreen, setIsFullscreen] = createSignal(false);
-  const { isAnimationOn, setIsAnimationOn } = useAnimationStore();
-
-  const toggleFullscreen = () => {
-    const canvas = document.getElementById('background');
-    if (!canvas) return;
-
-    if (!document.fullscreenElement) {
-      canvas.requestFullscreen().then(() => {
-        setIsFullscreen(true);
-        // Ensure animation is on when entering fullscreen
-        if (!isAnimationOn()) {
-          setIsAnimationOn(true);
-        }
-      }).catch(err => {
-        console.error(`Error attempting to enable fullscreen: ${err.message}`);
-      });
-    } else {
-      document.exitFullscreen().then(() => {
-        setIsFullscreen(false);
-      }).catch(err => {
-        console.error(`Error attempting to exit fullscreen: ${err.message}`);
-      });
-    }
-  };
-
-  // Listen for fullscreen change events
-  onMount(() => {
-    const handleFullscreenChange = () => {
-      setIsFullscreen(!!document.fullscreenElement);
-    };
-
-    document.addEventListener('fullscreenchange', handleFullscreenChange);
-
-    onCleanup(() => {
-      document.removeEventListener('fullscreenchange', handleFullscreenChange);
-    });
-  });
+  const { isFullscreen, toggleFullscreen } = useFullscreenStore();
   return (
     <>
       {/* Background. Mounting point for lava animation - only rendered on client */}
@@ -67,7 +31,7 @@ export default function App() {
 
       {/* Fullscreen button - desktop only */}
       <button
-        onClick={toggleFullscreen}
+        onClick={() => toggleFullscreen('background-container')}
         class="fixed top-4 right-4 z-[9999] p-2 text-white/0 hover:text-white/80 transition-opacity hidden md:block cursor-pointer"
         aria-label={isFullscreen() ? "Exit fullscreen" : "Enter fullscreen and start animation"}
         title={isFullscreen() ? "Exit fullscreen" : "Enter fullscreen and start animation"}
